turboquant-sys: drop unnecessary `as u8` cast (clippy::unnecessary_cast)
Rust 1.95's clippy adds `clippy::unnecessary_cast` to the `-D warnings`
set, which fired on the bit-pack write in `compress_values`:

    output[target_byte] |= (((code >> bit) & 1) << intra_byte) as u8;

Both operands of the shift are already `u8`, so the cast is a no-op.
Drop it.

The change is mechanical; all 7 turboquant-sys tests still pass and the
host's TurboQuant integration test (`turboquant_round_trip_through_native_rust_implementation`)
is unaffected.

Unblocks the Linux CI lint step that has been failing since `9f482d0`.

+++ turboquant-sys/src/lib.rs
@@ -112,7 +112,7 @@ pub fn compress_values(input: &[f32], bits: u8) -> Result<Vec<u8>, TurboQuantErr
             let intra_byte = target_bit % 8;
             // Each codebook is a power of two with at most 3 bits, so the AND-mask
             // and shift cannot overflow `u8`.
-            output[target_byte] |= (((code >> bit) & 1) << intra_byte) as u8;
+            output[target_byte] |= ((code >> bit) & 1) << intra_byte;
         }
         bit_index += bits as usize;
     }
